# backend/openapi.yaml
openapi: 3.1.0
info:
  title: Home Logbook API
  version: 0.11.0
servers:
- url: http://192.168.68.62:8787/api/v1
security:
- bearerAuth: []
paths:
...
  /event-types:
    get:
      summary: List event types with recent events
      parameters:
      - in: query
        name: recentLimit
        schema:
          type: integer
          minimum: 0
          maximum: 20
          default: 5
      - in: query
        name: includeInactive
        schema:
          type: boolean
          default: false
      responses:
        '200':
          description: Event type summaries
    post:
      summary: Create an event type
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventTypeCreate'
      responses:
        '201':
          description: Created
        '400':
          description: Invalid unit selection
      parameters:
      - $ref: '#/components/parameters/IdempotencyKey'
  /event-types/{key}:
    parameters:
    - in: path
      name: key
      required: true
      schema:
        type: string
    get:
      summary: Get one event type
      responses:
        '200':
          description: Event type
    patch:
      summary: Update or archive an event type
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventTypeUpdate'
      responses:
        '200':
          description: Updated
        '409':
          description: Historical numeric data locks the measurement dimension
      parameters:
      - $ref: '#/components/parameters/IdempotencyKey'
    delete:
      summary: Delete an unused event type
      responses:
        '204':
          description: Deleted
      parameters:
      - $ref: '#/components/parameters/IdempotencyKey'
  /event-types/{key}/latest-event:
    parameters:
    - in: path
      name: key
      required: true
      schema:
        type: string
    get:
      summary: Get the most recent event for an event type
      description: Uses the event type key and returns the newest event ordered by startedAt, then id.
      responses:
        '200':
          description: Most recent event
        '404':
          description: Unknown event type or no recorded events
    patch:
      summary: Update the most recent event for an event type
      description: Updates only supplied start-time, value, unit, text-value, note, or metadata fields. Numeric
        values are validated and normalized by the backend. For point events, changing startedAt also changes endedAt
        to the same timestamp. For duration events, the existing end time is preserved.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LatestEventUpdate'
      responses:
        '200':
          description: Updated most recent event
        '400':
          description: Invalid value
          unit: null
          or start time: null
        '404':
          description: Unknown event type or no recorded events
        '409':
          description: The selected latest event changed concurrently
      parameters:
      - $ref: '#/components/parameters/IdempotencyKey'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    EventTypeCreate:
      type: object
      required:
      - key
      - name
      properties:
        key:
          type: string
          pattern: ^[a-z][a-z0-9_-]{0,63}$
        name:
          type: string
        description:
          type:
          - string
          - 'null'
        unitTypeKey:
          type:
          - string
          - 'null'
          examples:
          - energy
        defaultUnitKey:
          type:
          - string
          - 'null'
          examples:
          - kwh
        icon:
          type:
          - string
          - 'null'
        color:
          type:
          - string
          - 'null'
          pattern: ^#[0-9A-Fa-f]{6}$
        voiceAliases:
          type: array
          maxItems: 50
          items:
            type: string
            minLength: 1
            maxLength: 120
    EventTypeUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type:
          - string
          - 'null'
        unitTypeKey:
          type:
          - string
          - 'null'
        defaultUnitKey:
          type:
          - string
          - 'null'
        icon:
          type:
          - string
          - 'null'
        color:
          type:
          - string
          - 'null'
        isActive:
          type: boolean
        voiceAliases:
          type: array
          maxItems: 50
          items:
            type: string
            minLength: 1
            maxLength: 120
    EventValue:
      type: object
      properties:
        value:
          type:
          - number
          - 'null'
          description: Original input value
        unitKey:
          type:
          - string
          - 'null'
          description: Exact compatible input unit
        textValue:
          type:
          - string
          - 'null'
        note:
          type:
          - string
          - 'null'
        metadata:
          type: object
          additionalProperties: true
    LatestEventUpdate:
      allOf:
      - type: object
        properties:
          startedAt:
            type: string
            format: date-time
      - $ref: '#/components/schemas/EventValue'
      minProperties: 1
      description: Supply at least one editable field. unitKey may only be supplied with a numeric value. Changing
        a point event's startedAt moves its endedAt to the same instant; a duration event keeps its current endedAt.
...
  parameters:
    IdempotencyKey:
      in: header
      name: Idempotency-Key
      required: false
      description: Optional unique key for safely retrying POST, PATCH, and DELETE requests. Reusing a key with
        the same request returns the stored response; reusing it with a different request returns 409.
      schema:
        type: string
        minLength: 1
        maxLength: 200